More options

+++ Install.sh
@@ -3,21 +3,31 @@
 echo "Enter URL you'd like the browser to open at boot:"
 read pageURL
 printf -v pageURL "%q\n" "$pageURL"
-echo "How often (in minutes) do you want the page to refresh? (0 to disable)"
-read refreshRate
 read -r -p "Hide cursor? (Good for non-interactive displays) [y/N] " hideCursor
-
+read -r -p "Clear browser cache at boot? [y/N] " hideCursor
+read -p "How often (in minutes) do you want the page to refresh? (0 to disable): " refreshRate
 while ! [[ $refreshRate =~ ^-?[0-9]+$ ]]
 do
     echo "Error: please enter an integer."
-    echo "How often (in minutes) do you want the page to refresh? (0 to disable)"
-    read refreshRate
+    read -p "How often (in minutes) do you want the page to refresh? (0 to disable): " refreshRate
+done
+
+read -p "How long to wait before launching browser (in seconds)? (0 to disable, 60-120 recommended so network has time to connect): " launchWait
+while ! [[ $launchWait =~ ^-?[0-9]+$ ]]
+do
+    echo "Error: please enter an integer."
+    read -p "How long to wait before launching browser (in seconds)? (0 to disable, 60-120 recommended so network has time to connect): " launchWait
 done
 
 HOMEDIR=~
 
+# Allow Pi time to boot
 echo "export DISPLAY=:0" > ~/rpi-webpage-display-atBoot.sh
-echo "rm -rf ~/.cache/chromium/*" >> ~/rpi-webpage-display-atBoot.sh
+echo "sleep $launchWait" >> ~/rpi-webpage-display-atBoot.sh
+if [[ "$clearCache" =~ ^([yY][eE][sS]|[yY])+$ ]]
+then
+    echo "rm -rf ~/.cache/chromium/*" >> ~/rpi-webpage-display-atBoot.sh
+fi
 echo "chromium-browser --no-first-run --disable --disable-translate --disable-infobars ----disable-session-crashed-bubble --disable-translate --disable-suggestions-service --disable-save-password-bubble --noerrdialogs --start-maximized --kiosk \"${pageURL}\" &" >> ~/rpi-webpage-display-atBoot.sh
 sudo chmod +x "${HOMEDIR}/rpi-webpage-display-atBoot.sh"
 (crontab -l 2>/dev/null; echo "@reboot ${HOMEDIR}/rpi-webpage-display-atBoot.sh") | crontab -
@@ -37,9 +47,6 @@ then
     echo "xdotool search --onlyvisible --class \"Chromium\" windowfocus key 'ctrl+r'" >> ~/rpi-webpage-display-Refresh.sh
     sudo chmod +x "${HOMEDIR}/rpi-webpage-display-Refresh.sh"
     (crontab -l 2>/dev/null; echo "*/${refreshRate} * * * * ${HOMEDIR}/rpi-webpage-display-Refresh.sh") | crontab -
-
-    # Refresh page after boot in case networking wasn't fully up or other issue
-    echo "sleep 120 && xdotool search --onlyvisible --class \"Chromium\" windowfocus key 'ctrl+r'" >> ~/rpi-webpage-display-atBoot.sh
 fi
 
 echo "Done! Rebooting now..."
